test: 잘못된 테스트 로직 수정
