+++ src/main/java/chess/domain/chessboard/attribute/Rank.java
@@ -7,8 +7,8 @@ public enum Rank {
 
     EIGHT, SEVEN, SIX, FIVE, FOUR, THREE, TWO, ONE;
 
-    private static final int RANK_MAX = 8;
     private static final int RANK_MIN = 1;
+    private static final int RANK_MAX = 8;
 
     public static Rank from(final char value) {
         return from(String.valueOf(value));
@@ -31,7 +31,7 @@ public enum Rank {
     }
 
     public static boolean isInRange(final int column) {
-        return RANK_MIN <= column && column <= RANK_MAX;
+        return EIGHT.toRow() <= column && column <= ONE.toRow();
     }
 
     public int toRow() {
